requirements.txt
# Zero-Inflated Count Model Pipeline Requirements
# ======================================
# Pinned versions for reproducibility - tested on Apple M3 Pro

# Core data processing
numpy==1.24.3
pandas==2.0.2
scipy==1.10.1

# Machine learning
scikit-learn==1.3.0
xgboost==1.7.5
lightgbm==4.0.0
catboost==1.2.2
statsmodels==0.14.0

# Deep learning with Apple Metal support
tensorflow==2.13.0
tensorflow-metal==1.0.1

# Visualization
matplotlib==3.7.1
seaborn==0.12.2
shap==0.42.1

# Hyperparameter optimization
optuna==3.2.0

# Experiment tracking
mlflow==2.4.1

# Model serialization
joblib==1.3.1

# Configuration
pyyaml==6.0.1
pydantic==1.10.8

# Development tools
pytest==7.4.0
pytest-cov==4.1.0
line_profiler==4.0.3
black==23.3.0
flake8==6.0.0
isort==5.12.0
mypy==1.4.1

# Jupyter for reporting
jupyter==1.0.0
jupyterlab==4.0.2
nbconvert==7.6.0

# Hardware requirements (for documentation):
# - Apple M3 Pro or better
# - 16GB+ RAM recommended for 15M dataset
# - macOS 13.0+ for TensorFlow Metal support
